+++ src/BackpackServiceProvider.php
@@ -18,8 +18,7 @@ class BackpackServiceProvider extends ServiceProvider
         \Backpack\CRUD\app\Console\Commands\CreateUser::class,
         \Backpack\CRUD\app\Console\Commands\PublishBackpackUserModel::class,
         \Backpack\CRUD\app\Console\Commands\PublishBackpackMiddleware::class,
-        \Backpack\CRUD\app\Console\Commands\InstallCrud::class,
-        \Backpack\CRUD\app\Console\Commands\PublishCrud::class,
+        \Backpack\CRUD\app\Console\Commands\PublishView::class,
     ];
 
     // Indicates if loading of the provider is deferred.